requirements.txt
# external requirements
importlib-metadata<5.0
importlib-resources==5.12.0
flake8
python-dotenv>=0.5.1
pytest
numpy
pandas>=1.3.0
seaborn
matplotlib
scipy
pytorch-lightning==1.4.5
torchmetrics==0.6.0
hydra-core
hydra_colorlog
mlflow
scikit-learn
tables
ray[tune]

git+https://github.com/fadel/pytorch_ema
